pyproject.toml
[project]
name = "perso-qa-computer-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
